Improve TarFileSystem once more
Along the lines, refactor more common code from ZipFileSystem and
LibArchiveFileSystem into AbstractArchiveFileSystem.

Make the "test_isdir_isfile" test work.
