+++ fsspec/implementations/libarchive.py
@@ -14,7 +14,7 @@ from ctypes import (
 import libarchive
 import libarchive.ffi as ffi
 
-from fsspec import AbstractArchiveFileSystem, AbstractFileSystem, open_files
+from fsspec import AbstractArchiveFileSystem, open_files
 from fsspec.implementations.memory import MemoryFile
 from fsspec.utils import DEFAULT_BLOCK_SIZE
 
@@ -65,7 +65,7 @@ def custom_reader(file, format_name="all", filter_name="all", block_size=ffi.pag
         yield libarchive.read.ArchiveRead(archive_p)
 
 
-class LibArchiveFileSystem(AbstractFileSystem, AbstractArchiveFileSystem):
+class LibArchiveFileSystem(AbstractArchiveFileSystem):
     """Compressed archives as a file-system (read-only)
 
     Supports the following formats:
@@ -177,43 +177,6 @@ class LibArchiveFileSystem(AbstractFileSystem, AbstractArchiveFileSystem):
                 for dirname in self._all_dirnames(list_names)
             }
         )
-
-    def info(self, path, **kwargs):
-        self._get_dirs()
-        path = self._strip_protocol(path)
-        if path in self.dir_cache:
-            return self.dir_cache[path]
-        elif path + "/" in self.dir_cache:
-            return self.dir_cache[path + "/"]
-        else:
-            raise FileNotFoundError(path)
-
-    def ls(self, path, detail=False, **kwargs):
-        self._get_dirs()
-        paths = {}
-
-        for p, f in self.dir_cache.items():
-            p = p.rstrip("/")
-            if "/" in p:
-                root = p.rsplit("/", 1)[0]
-            else:
-                root = ""
-            if root == path.rstrip("/"):
-                paths[p] = f
-            elif all(
-                (a == b)
-                for a, b in zip(path.split("/"), [""] + p.strip("/").split("/"))
-            ):
-                # root directory entry
-                ppath = p.rstrip("/").split("/", 1)[0]
-                if ppath not in paths:
-                    out = {"name": ppath + "/", "size": 0, "type": "directory"}
-                    paths[ppath] = out
-        out = list(paths.values())
-        if detail:
-            return out
-        else:
-            return list(sorted(f["name"] for f in out))
 
     def _open(
         self,
