+++ fsspec/implementations/tar.py
@@ -12,7 +12,7 @@ typemap = {b"0": "file", b"5": "directory"}
 logger = logging.getLogger("tar")
 
 
-class TarFileSystem(fsspec.AbstractFileSystem, AbstractArchiveFileSystem):
+class TarFileSystem(AbstractArchiveFileSystem):
     def __init__(
         self, fo, index_store=None, storage_options=None, compression=None, **kwargs
     ):
@@ -34,45 +34,35 @@ class TarFileSystem(fsspec.AbstractFileSystem, AbstractArchiveFileSystem):
             fo = compr[compression](fo)
 
         self.fo = fo
+        self.tar = tarfile.TarFile(fileobj=self.fo)
+        self.dir_cache = None
+
         self.index_store = index_store
         self.index = None
         self._index()
 
     def _index(self):
         # TODO: load and set saved index, if exists
-        fo = tarfile.TarFile(fileobj=self.fo)
         out = {}
-        for ti in fo:
+        for ti in self.tar:
             info = ti.get_info()
             info["type"] = typemap[info["type"]]
-            out[ti.get_info()["name"]] = (info, ti.offset_data)
+            name = ti.get_info()["name"].rstrip("/")
+            out[name] = (info, ti.offset_data)
 
         self.index = out
         # TODO: save index to self.index_store here, if set
 
-    def ls(self, path, detail=False, **kwargs):
-        path = self._strip_protocol(path)
-        parts = path.rstrip("/").split("/")
-        if parts and parts[0] == "":
-            parts = []
-        out = []
-        for name, (details, _) in self.index.items():
-            nparts = name.rstrip("/").split("/")
-            if parts == nparts and details["type"] != "directory":
-                out = [details]
-                break
-            if len(nparts) and parts == nparts[:-1]:
-                out.append(details)
-        if detail:
-            return out
-        else:
-            return [o["name"] for o in out]
+    def _get_dirs(self):
 
-    def info(self, path, **kwargs):
-        try:
-            return self.index[path][0]
-        except KeyError:
-            raise FileNotFoundError(path)
+        if self.dir_cache is not None:
+            return
+
+        self.dir_cache = {}
+        for member in self.tar.getmembers():
+            info = member.get_info()
+            info["type"] = typemap[info["type"]]
+            self.dir_cache[info["name"]] = info
 
     def _open(self, path, mode="rb", **kwargs):
         if mode != "rb":
