+++ fsspec/spec.py
@@ -1568,7 +1568,7 @@ class AbstractBufferedFile(io.IOBase):
         self.close()
 
 
-class AbstractArchiveFileSystem:
+class AbstractArchiveFileSystem(AbstractFileSystem):
     def ukey(self, path):
         return tokenize(path, self.fo, self.protocol)
 
@@ -1584,3 +1584,39 @@ class AbstractArchiveFileSystem:
 
         dirnames = {self._parent(path) for path in paths} - {self.root_marker}
         return dirnames | self._all_dirnames(dirnames)
+
+    def info(self, path, **kwargs):
+        self._get_dirs()
+        path = self._strip_protocol(path)
+        if path in self.dir_cache:
+            return self.dir_cache[path]
+        elif path + "/" in self.dir_cache:
+            return self.dir_cache[path + "/"]
+        else:
+            raise FileNotFoundError(path)
+
+    def ls(self, path, detail=False, **kwargs):
+        self._get_dirs()
+        paths = {}
+        for p, f in self.dir_cache.items():
+            p = p.rstrip("/")
+            if "/" in p:
+                root = p.rsplit("/", 1)[0]
+            else:
+                root = ""
+            if root == path.rstrip("/"):
+                paths[p] = f
+            elif all(
+                (a == b)
+                for a, b in zip(path.split("/"), [""] + p.strip("/").split("/"))
+            ):
+                # root directory entry
+                ppath = p.rstrip("/").split("/", 1)[0]
+                if ppath not in paths:
+                    out = {"name": ppath + "/", "size": 0, "type": "directory"}
+                    paths[ppath] = out
+        out = list(paths.values())
+        if detail:
+            return out
+        else:
+            return list(sorted(f["name"] for f in out))
